backend/requirements.txt
chromadb>=1.0.13
fastapi>=0.115.14
langchain>=0.3.26
langchain-chroma>=0.2.4
langchain-openai>=0.3.27
mem0ai>=0.1.112
nltk>=3.9.1
rank-bm25>=0.2.2
python-dotenv>=1.0.0
websockets>=12.0
onnxruntime>=1.22.0
pydantic-settings>=2.0.0
pytest>=8.4.1